Controls select, deselect, and move as desired
No testing has been done to prove reliability in edge cases.

Controls highlight routes when chosen, fill in input when routes are
chosen, and remove input when routes are deselected. All as desired.

Functionality with Ticket panel is unknown. Controls is not nearly as
modular as I would like it. We may be firing into Tickets from a wide
variety of places.