DB re-seeded
There was an error in the way flights were generated, resulting in
only one airline per route. (A second airline would overwrite
the first, etc.) Therefore the flights were re-seeded.

The reason for this error was that the script was trying to avoid
duplicate routes. The script builds both directions at one time,
and the prior check was trying to avoid building a route again if
the other direction had already taken care of things. By taking out
this check there are now multiple airlines on a single route, but now
there are far too many airline routes. There can be 50 flights a
day between airports in a single direction.

While a real issue, the database is not being reseeded again for
issue. It's too minor when it comes down to the presentation.

Old flights remain in database, but the airport references were
updated.